+++ backend/database/seed.go
@@ -304,7 +304,7 @@ func seedService(model *models.Model) {
 			AccessKey:          os.Getenv("ACCESS_KEY_PFL"),
 			Token:              os.Getenv("TOKEN_PFL"),
 			Timestamp:          os.Getenv("TIMESTAMP_PFL"),
-			ShortDescription:   "Liveness detection is a ML based solution to determine the likelihood of an organization interfacing with a physically present human being and an inanimate spoof artifact.",
+			ShortDescription:   "Liveness detection is a Machine Learning based solution to check the selfie is presented a human being or an inanimate spoof artifact.",
 			LongDescription:    "Liveness detection is a Machine Learning based solution to determine the likelihood of an organization interfacing with a physically present human being and an inanimate spoof artifact.",
 			SpecialInstruction: "Special Instruction for Passive Face Liveness Detection",
 			CreatedAt:          time.Now(),