+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
 ## [Unreleased]
 ### Changed
 
 - [FRONTEND] Modify analytic icon to use icon.png name instead of analytic thumbnail field
+- [BACKEND] Add company size field
+
 ## [1.2.1 - 13/01/2022]
+
 ### Changed
 
 - [FRONTEND] Fix solution card external link
+
 ## [1.2.0 - 13/01/2022]
 
 ### Added
@@ -21,7 +26,6 @@ All notable changes to this project will be documented in this file.
 - [FRONTEND] Update solution card design
 - [FRONTEND] Change NIST logo to be smaller in larger device
 - [FRONTEND] Add Partner Solutions Carousel
-
 
 ### Removed