+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
 ### Changed
+
+### Removed
+
+## [1.4.0 - 27/01/2022]
 
 ### Added